+++ src/main/java/org/example/dao/UserDao.java
@@ -8,7 +8,7 @@ public interface UserDao {
 
 
     public List<User> index();
-    public User show (int id);
+    public User show (Long id);
     public void save (User user);
     public void update (User user);
     public void delete (User user);
