+++ src/main/java/org/example/service/UserService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface UserService {
     public List<User> index();
-    public User show (int id);
+    public User show (Long id);
     public void save (User user);
     public void update (User user);
     public void delete (User user);
